Ensure get methods always return a string array
If only one item exists in HiddenSettings or ShownSettings, the return statement unwinds the array and returns a single string. This would make calling the += operator concatenate the string passed to it when only one setting was present. Making it always return an array ensures that the += operator works as intended.

https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_return#return-values-and-the-pipeline